--- src/main/java/com/example/beans/Brush.java
@@ -1,15 +1,0 @@
-package com.example.beans;
-
-public class Brush {
-    String color;
-    public boolean fill;
-    public String fillColor;
-    public double transparency;
-    public Points points;
-    public double thickness;
-
-    public class Points {
-        public double[][] x;
-        public double[][] y;
-    }
-}
