--- src/main/java/com/example/beans/Waypoint.java
@@ -1,11 +1,0 @@
-package com.example.beans;
-
-public class Waypoint {
-    public double x;
-    public double y;
-    public String content;
-    public String name;
-    public int width;
-    public int height;
-    public boolean active;
-}
