+++ src/main/java/com/mapper/beans/Waypoint.java
@@ -1,0 +1,11 @@
+package com.mapper.beans;
+
+public class Waypoint {
+    public double x;
+    public double y;
+    public String content;
+    public String name;
+    public int width;
+    public int height;
+    public boolean active;
+}
