--- src/test/java/com/example/servingwebcontent/ServingWebContentApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.servingwebcontent;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ServingWebContentApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
