+++ src/test/java/com/mapper/servingwebcontent/ServingWebContentApplicationTests.java
@@ -1,0 +1,13 @@
+package com.mapper.servingwebcontent;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServingWebContentApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
